requirements.txt
requests>=2.28.0
beautifulsoup4>=4.11.0
sentence-transformers>=2.2.0
chromadb>=0.4.0
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
lxml>=4.9.0
pydantic>=2.0.0
openai>=1.0.0
numpy>=1.24.0
torch>=2.0.0